+++ src/nn/conv.rs
@@ -1,4 +1,4 @@
-use crate::gradients::{CanUpdateWithGradients, GradientProvider, Tape, UnusedTensors};
+use crate::gradients::*;
 use crate::prelude::*;
 use rand::Rng;
 use rand_distr::Uniform;
@@ -55,6 +55,7 @@ impl<const I: usize, const O: usize, const K: usize, const S: usize, const P: us
     }
 }
 
+#[cfg(feature = "nightly")]
 impl<
         T: Tape,
         const I: usize,
@@ -76,6 +77,7 @@ where
     }
 }
 
+#[cfg(feature = "nightly")]
 impl<
         T: Tape,
         const B: usize,
@@ -109,6 +111,7 @@ where
     }
 }
 
+#[cfg(feature = "nightly")]
 #[cfg(test)]
 mod tests {
     use super::*;
