+++ src/nn/transformer/encoder.rs
@@ -103,6 +103,7 @@ where
     }
 }
 
+#[cfg(feature = "nightly")]
 #[cfg(test)]
 mod tests {
     use super::*;
